Luciferase-x5i.6: extract KnnSearcher (RDR-008 P3-main)
P3-main of the locked AbstractSpatialIndex decomposition. Moves the k-NN
cluster out of the 5409-line façade into a new lucien.cache.KnnSearcher
feature object implementing the KnnProvider sub-interface the P3-prep
split introduced.

Cluster moved:
- public entry kNearestNeighbors (now a façade delegator)
- 7 private search helpers: performKNNExpandingRadiusSearch,
  performKNNSFCBasedSearch, performKNNSFCRangePruning + Morton/Tetree
  variants, performFullScanKNN, searchKNNInRadius, findNearestNodeToPoint
- 4 AtomicLong performance counters
- KNNPerformanceMetrics record (was a public inner record on the façade;
  the only external caller binds via var, so the move is non-breaking)

KnnGeometry sub-interface grew from 4 methods to 10 — added the spatial-
cluster hooks the moved code consumes (getCellSizeAtLevel,
findNodesIntersectingBounds, knnRequiresFullDomainSweep,
addNeighboringNodes), the entity-cache lookup (getCachedEntityPosition,
duplicated with FrustumGeometry by intent until a third consumer
emerges), and the façade depth constant (maxDepth).

Façade integration:
- Drops 'implements KnnProvider<Key,ID>' from the class declaration
  (P3-prep added it transiently as a scaffold; P3-main moves the role
  to KnnSearcher itself).
- GhostCoordinator now takes the KnnSearcher instance as its KnnProvider
  argument instead of the partially-constructed façade.
- ctor wires 'this.knn = new KnnSearcher<>(core, new KnnGeometryImpl(),
  () -> this.lockingStrategy)' before the ghost ctor so the latter has
  a KnnProvider to consume. The lockingStrategy supplier is late-bound
  because configureFineGrainedLocking can replace the field.

Concurrency invariant preserved verbatim (RDR-008 §Open Questions §2):
the version-read-outside-lock pattern stays — kNearestNeighbors reads
core.spatialVersion().get() before entering the locked region and writes
the cache result under that snapshot. KNNCache.get returns null on
version mismatch, so a concurrent updateEntity between version-read and
cache-write cannot return stale results; at worst the next reader sees
a miss. No new synchronization required.

Defensive DRY: the three byte-for-byte-identical inner candidate loops
(Morton / Tetree / FullScan) consolidate into a shared
collectCandidatesFromNode helper. Behaviorally equivalent (continue
guards match the original nested-if structure).

volatile lockingStrategy: the substantive-critic flagged a JMM
publication gap — configureFineGrainedLocking writes the field under
write-lock, but KnnSearcher's supplier reads it outside any lock from a
different package. Adding volatile closes the gap (and mirrors the dsoc
field which is already volatile for the identical pattern). Pre-extraction
the same field was read outside any lock from inside the façade; the
gap was latent, the extraction makes it cross-package and visible.

Cumulative façade arc since P0 start:
  AbstractSpatialIndex.java: 5851 → 4916 lines (-935 cumulative; -493
  for P3-main). 4 of 6 phases done (DSOC, distributed-ghost, k-NN, plus
  the P3-prep sub-interface split). G6 target ≈70 methods remaining.

Subclasses (Octree, Tetree, Prism, SFCArrayIndex) unchanged (P1-P4
invariant). git diff --stat shows only AbstractSpatialIndex.java +
KnnGeometry.java + new KnnSearcher.java touched.

Tests: 2441, 0 failures, 0 errors, 22 skipped — identical to baseline.

Refs: RDR-008 §Decision item 1 (with P3 refinement), bead Luciferase-x5i.6.
